Se realiza prueba de 30 trials por cada clase. Se modifica levemente getData de EEGLogger para obtener muestras entereas. Todo OK. Se puede continuar con la siguiente fase correspondiente a la sesión de Calibración/Feedback para probar pipeline de clasificación

+++ Desarrollo/PythonScripts/scripts/EEGLogger/eegLogger.py
@@ -44,7 +44,7 @@ class EEGLogger():
         Los datos se entregan en un numpy array de forma [canales, muestras]. Los datos están en microvolts.
         - sampleLength: duración (en segundos) de la señal a adquirir de la placa. Por defecto son 6 segundos."""
 
-        num_samples = self.board.get_sampling_rate(self.board_id) * sampleLength
+        num_samples = int(self.board.get_sampling_rate(self.board_id) * sampleLength)
         return self.board.get_board_data(num_samples)
     
     def addData(self, newdata):
